requirements.txt
accelerate==0.25.0
matplotlib==3.8.0
numpy==1.25.2
opencv_python==4.8.1.78
pandas==2.0.3
PyYAML==6.0.2
scipy==1.14.1
setuptools==68.0.0
torch==2.2.1
tqdm==4.66.1
